+++ js/runtime-config.js
@@ -129,7 +129,7 @@
   if (!siteBasePath.endsWith('/')) siteBasePath += '/';
 
   var config = Object.freeze({
-    version: '20260411-f22d',
+    version: '20260411-f23a',
     site: Object.freeze({
       siteBaseUrl: siteBaseUrl,
       siteBasePath: siteBasePath,
@@ -138,7 +138,7 @@
       isSpecialUrl: isSpecialUrl
     }),
     sw: Object.freeze({
-      version: '20260411-f22d',
+      version: '20260411-f23a',
       scriptFilename: 'sw.js',
       messages: SW_MESSAGES
     }),